pos/backend/scripts/requirements.txt
google-cloud-firestore==2.16.0
google-auth==2.34.0
